requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
et_xmlfile==2.0.0
greenlet==3.2.4
idna==3.11
loguru==0.7.3
numpy==2.3.5
openpyxl==3.1.5
pandas==2.3.3
playwright==1.56.0
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
win32_setctime==1.2.0
